refact: modularize and clean common
Split scr/common into focused modules: integrator.py keeps only the
public integrate() / make_events() API, _solvers.py holds the
DOP853/RK45/Verlet backends with Brent event refinement, and
_numba_kernels.py holds the numba-jitted hot paths used by parallel.py.
parallel.py updated accordingly.

Also drops the stale scr/black_holes/kerr_mog.pdf (not source code).
